Title: 세 수의 최대 곱, Time: -1ms, Memory: -1MB, Status: Compile Error - CodeTree

+++ 240530/세 수의 최대 곱/maximum-product-of-three-numbers.cpp
@@ -10,7 +10,7 @@ int main() {
     {
         cin>>arr[i];
     }
-    int maxi=0;
+    long long maxi=0;
     for(int i=0;i<n-2;i++)
     {
         for(int j=i+1;j<n-1;j++)
